Split encoders into multiple files
